Feature : contenido de categorias de Descripcion de Curso en TABS - Refs #10529

+++ main/course_description/index.php
@@ -51,15 +51,6 @@ if (isset($_GET['isStudentView']) && $_GET['isStudentView'] == 'true') {
 
 // interbreadcrumb
 $interbreadcrumb[] = array ("url" => "index.php", "name" => get_lang('CourseProgram'));
-if(intval($description_type) == 1) $interbreadcrumb[] = array ("url" => "#", "name" => get_lang('GeneralDescription'));
-if(intval($description_type) == 2) $interbreadcrumb[] = array ("url" => "#", "name" => get_lang('Objectives'));
-if(intval($description_type) == 3) $interbreadcrumb[] = array ("url" => "#", "name" => get_lang('Topics'));
-if(intval($description_type) == 4) $interbreadcrumb[] = array ("url" => "#", "name" => get_lang('Methodology'));
-if(intval($description_type) == 5) $interbreadcrumb[] = array ("url" => "#", "name" => get_lang('CourseMaterial'));
-if(intval($description_type) == 6) $interbreadcrumb[] = array ("url" => "#", "name" => get_lang('HumanAndTechnicalResources'));
-if(intval($description_type) == 7) $interbreadcrumb[] = array ("url" => "#", "name" => get_lang('Assessment'));
-if(intval($description_type) == 8) $interbreadcrumb[] = array ("url" => "#", "name" => get_lang('ThematicAdvance'));
-if(intval($description_type) >= 9) $interbreadcrumb[] = array ("url" => "#", "name" => get_lang('Others'));
 
 // course description controller object
 $course_description_controller = new CourseDescriptionController();
@@ -72,20 +63,32 @@ switch ($action) {
 	case 'history':		
 		$course_description_controller->listing(true);
 		break;
-	case 'add'	  :   
-		if (api_is_allowed_to_edit(null,true)) {
-			$course_description_controller->add();
-		}
+	case 'add'	  :
+        if (api_is_platform_admin()) {
+            if (api_is_allowed_to_edit(null,true)) {
+                $course_description_controller->add();
+            }
+        } else {
+            header('Location: '.$_SERVER['PHP_SELF']);
+        }
 		break;
 	case 'edit'	  :	
-		if (api_is_allowed_to_edit(null,true)) {
-			$course_description_controller->edit($id, $description_type);
-		}
+		if (api_is_platform_admin()) {
+            if (api_is_allowed_to_edit(null,true)) {
+                $course_description_controller->edit($id, $description_type);
+            }
+        } else {
+            header('Location: '.$_SERVER['PHP_SELF']);
+        }
 		break;
 	case 'delete' :	
-		if (api_is_allowed_to_edit(null,true)) {
-			$course_description_controller->destroy($id);
-		}
+        if (api_is_platform_admin()) {
+            if (api_is_allowed_to_edit(null,true)) {
+                $course_description_controller->destroy($id);
+            }
+        } else {
+            header('Location: '.$_SERVER['PHP_SELF']);
+        }
 		break;
 	default		  :	
 		$course_description_controller->listing();
